refactor: update user repository to use UserService interface; add UserController and DTOs for user management
