--- src/main/java/com/develop/users_microservice/domain/service/PasswordEncoder.java
@@ -1,6 +1,0 @@
-package com.develop.users_microservice.domain.service;
-
-public interface PasswordEncoder {
-    String encode(String rawPassword);
-    boolean matches(String rawPassword, String encodedPassword);
-}
