+++ src/main/java/com/develop/users_microservice/infrastructure/repository/UserRepositoryImpl.java
@@ -1,9 +1,8 @@
 package com.develop.users_microservice.infrastructure.repository;
 import com.develop.users_microservice.domain.model.Role;
 import com.develop.users_microservice.domain.model.User;
-import com.develop.users_microservice.domain.repository.UserRepository;
+import com.develop.users_microservice.domain.interfaces.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -11,7 +10,7 @@ import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
-public class UserRepositoryImpl implements UserRepository{
+public class UserRepositoryImpl implements UserService {
 
     private final JpaUserRepository jpaUserRepository;
