--- src/main/java/com/develop/users_microservice/infrastructure/security/BCryptPasswordEncoderImpl.java
@@ -1,21 +1,0 @@
-package com.develop.users_microservice.infrastructure.security;
-
-import com.develop.users_microservice.domain.service.PasswordEncoder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class BCryptPasswordEncoderImpl implements PasswordEncoder {
-    private final BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
-
-    @Override
-    public String encode(String rawPassword) {
-        return bCryptPasswordEncoder.encode(rawPassword);
-    }
-
-    @Override
-    public boolean matches(String rawPassword, String encodedPassword) {
-        return bCryptPasswordEncoder.matches(rawPassword, encodedPassword);
-    }
-}
-
