+++ src/main/java/com/develop/users_microservice/presentation/dto/UserResponseDTO.java
@@ -1,4 +1,4 @@
-package com.develop.users_microservice.application.dto;
+package com.develop.users_microservice.presentation.dto;
 import lombok.Getter;
 import lombok.Setter;
